Several changes; see description
Added a script to count the numbers of supernova types.
Calculated the supernova rate just on the inside of spiral galaxies.
Estimated the number of missing SNe due to the narrow KAIT field.
Added rate calculating functions which calculate rates normalized by
luminosity.
Added a script to test integrations of the Schechter function.
Added a K-luminosity attribute to the Galaxy object.